Add files via upload
In this project, we conduct a comprehensive analysis of customer reviews for electronic products sold on Amazon, utilizing a dataset of 30,000 reviews. Our objective is to apply various Natural Language Processing (NLP) techniques to extract meaningful insights and develop a model that predicts the sentiment of future reviews based on historical data. Through this analysis, we aim to understand customer sentiments and enhance decision-making for both consumers and businesses.